In this commit I have roughly sketched out the model, viewmodel and view for this application.
The main ContentView has a title, a space for the video player and a VideoDetailsView, the scrollable area containing the text. Once we have a video player, we will also need an overlay view to handle the custom controls.

The meat of the functionality will be inside the VideoPlayerViewModel. This object is responsbile for managing the server data, handling the AVPlayer and related setup and informing the view about things like the state of the player or whether or not the next/prev buttons are enabled. This viewmodel is added to the content view as an EnvironmentObject so it is accessible by all the views.