I am now playing the first video from the server in a VideoPlayer in the main screen. Had to add @MainActor to my ViewModel so it can make changes to it's published values on the main thread.
The VideoDetails view is now using live data as well, but I have not yet implemented the markdown so it is looking better but not great.

Coming soon, disabling the VideoPlayer controls and adding my own using the PlayerOverlayView, which also needs to use the pdf assets instead of placeholder text.