Improved the video overlay greatly in this commit. The next/previous buttons now work correctly, disabled if there are no videos to reach. The controls now also hide and reveal themself when the user taps on the video portion of the screen, so we don't have the controls showing on top of our video constantly.
There is a small problem with the state of the Play/Pause image. I have a bug where the wrong icon is being show; play whle the video is playing and pause while it is not, but the initial state is correct.

However, I am running out time, having started the work around 1:30 and now it is 5:00pm. I have not implemented Markdown, having tried a few attributed strings solutions I found on the internet that did not work properly. So my plan is to fix the markdown and then come back and deal with the button image bug.