Last bugfixes
I put a fair amount of effort into the error conditions that could occur in the viewmodel, in this commit I added the alert to the contentview so they would actually be displayed. I tested this by stopping the server and running the app; voila, alerts.

For a final touch, I felt the title font was much too large in the video details view so I changed that to headline.

This is the version of the exercise I will be submitting. Thank you to everyone at SilverOrange for the tryout and I hope to speak to you soon.